requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.2.2
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
mccabe==0.7.0
multidict==6.0.4
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyocr==0.8.3
requests==2.28.2
urllib3==1.26.15
yarl==1.8.2
